requirements.txt
crewai
crewai[tools]
sentence_transformers
PyPDF2
python-docx